refactor : error 메시지 상수 클래스 명 변경
error 메시지 상수 클래스 명 변경
