+++ src/main/java/lotto/exception/ConsumerException.java
@@ -1,15 +1,15 @@
 package lotto.exception;
 
-import static lotto.constants.ErrorConstants.ERROR_PREFIX;
+import static lotto.constants.ErrorMessage.ERROR_PREFIX;
 
-import lotto.constants.ErrorConstants;
+import lotto.constants.ErrorMessage;
 
 public class ConsumerException extends IllegalArgumentException {
     private ConsumerException(String message) {
         super(ERROR_PREFIX.getMessage() + message);
     }
 
-    public static ConsumerException of(ErrorConstants errorConstant) {
+    public static ConsumerException of(ErrorMessage errorConstant) {
         return new ConsumerException(errorConstant.getMessage());
     }
 }
