+++ src/main/java/lotto/exception/LottoException.java
@@ -1,15 +1,15 @@
 package lotto.exception;
 
-import static lotto.constants.ErrorConstants.ERROR_PREFIX;
+import static lotto.constants.ErrorMessage.ERROR_PREFIX;
 
-import lotto.constants.ErrorConstants;
+import lotto.constants.ErrorMessage;
 
 public class LottoException extends IllegalArgumentException {
     private LottoException(String message) {
         super(ERROR_PREFIX.getMessage() + message);
     }
 
-    public static LottoException of(ErrorConstants errorConstant) {
+    public static LottoException of(ErrorMessage errorConstant) {
         return new LottoException(errorConstant.getMessage());
     }
 }
